Seedling image classification
You are provided with a training set and a test set of images of plant
seedlings at various stages of grown. Each image has a filename that is its
unique id. The dataset comprises 12 plant species. The goal of the
competition is to create a classifier capable of determining a plant's
species from a photo. The list of species is as follows:
 Black-grass
 Charlock
 Cleavers
 Common Chickweed
 Common wheat
 Fat Hen
 Loose Silky-bent
 Maize
 Scentless Mayweed
 Shepherds Purse
 Small-flowered Cranesbill
 Sugar beet